feat: 노말빼빼로 만들기 함수 추가
