+++ Week1/week1_restruct/week1_restruct/main.swift
@@ -1,35 +1,24 @@
 import Foundation
 
-let bodyComponent: String = "*"
-let spaceForBarLine: String = " "
-let spaceForBarLineLength: Int = 4
-let bodyLineLength: Int = 11
+var peperoBodyComponent = "***"
+var peperoToppingComponent = "&"
+let peperoBarComponent = " | |"
 
-func drawIceCreamBody(bodyLength: Int) {
-    for _ in 1...bodyLength {
-        drawIceCreamBodyLine()
+func makeNormalPeperoBody(length: Int) {
+    for _ in 1...length {
+        print(" " + peperoBodyComponent)
     }
 }
 
-func drawIceCreamBodyLine() {
-    for _ in 1...bodyLineLength {
-        print(bodyComponent, terminator: "")
-    }
-    print("")
-}
-
-func drawIceCreamBar(barLength: Int) {
-    for _ in 1...barLength {
-        for _ in 1...spaceForBarLineLength {
-            print(spaceForBarLine,terminator: "")
-        }
-        print("| |")
+func makePeperoBar(length: Int) {
+    for _ in 1...length {
+        print(peperoBarComponent)
     }
 }
 
-func drawIceCream(bodyLength: Int, barLength: Int) {
-    drawIceCreamBody(bodyLength: bodyLength)
-    drawIceCreamBar(barLength: barLength)
+func makeNormalPepero(bodyLength: Int, barLength: Int) {
+    makeNormalPeperoBody(length: bodyLength)
+    makePeperoBar(length: barLength)
 }
 
-drawIceCream(bodyLength: 8, barLength: 4)
+
